refactor(refund): generalize RefundDialog over abstract RefundSource
